More wip, almost done writing everything to sentinel pairs

+++ include/Lz/detail/iterables/drop.hpp
@@ -28,7 +28,7 @@ public:
 
     template<class I = Iterable>
     LZ_NODISCARD LZ_CONSTEXPR_CXX_14 enable_if<sized<I>::value, std::size_t> size() const {
-        return lz::size(_iterable) - static_cast<std::size_t>(_n);
+        return static_cast<std::size_t>(lz::size(_iterable) - static_cast<std::size_t>(_n));
     }
 
     LZ_NODISCARD LZ_CONSTEXPR_CXX_14 iterator begin() && {
